@@ -1,5 +1,5 @@
 module github.com/erock530/gosf-socketio
 
-go 1.18
+go 1.22.0
 
-require github.com/gorilla/websocket v1.5.0
+require github.com/gorilla/websocket v1.5.3
