@@ -1,38 +1,14 @@
 package protocol
 
 const (
-	/**
-	Message with connection options
-	*/
 	MessageTypeOpen = iota
-	/**
-	Close connection and destroy all handle routines
-	*/
-	MessageTypeClose = iota
-	/**
-	Ping request message
-	*/
-	MessageTypePing = iota
-	/**
-	Pong response message
-	*/
-	MessageTypePong = iota
-	/**
-	Empty message
-	*/
-	MessageTypeEmpty = iota
-	/**
-	Emit request, no response
-	*/
-	MessageTypeEmit = iota
-	/**
-	Emit request, wait for response (ack)
-	*/
-	MessageTypeAckRequest = iota
-	/**
-	ack response
-	*/
-	MessageTypeAckResponse = iota
+	MessageTypeClose
+	MessageTypePing
+	MessageTypePong
+	MessageTypeEmpty
+	MessageTypeEmit
+	MessageTypeAckRequest
+	MessageTypeAckResponse
 )
 
 type Message struct {
@@ -42,4 +18,3 @@ type Message struct {
 	Args   string
 	Source string
 }
-
